refactor(react-surface): make plugins closures
