+++ packages/apps/plugins/plugin-markdown/src/components/index.ts
@@ -2,5 +2,7 @@
 // Copyright 2023 DXOS.org
 //
 
+export * from './MarkdownMainEmbedded';
+export * from './MarkdownMainEmpty';
 export * from './MarkdownMain';
 export * from './StandaloneLayout';
