+++ packages/apps/plugins/plugin-markdown/tsconfig.json
@@ -37,6 +37,9 @@
     },
     {
       "path": "../../../ui/aurora-theme"
+    },
+    {
+      "path": "../plugin-theme"
     }
   ]
 }
