+++ packages/apps/plugins/plugin-splitview/src/components/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export * from './SplitView';
+export * from './SplitViewMainContentEmpty';
