+++ packages/apps/plugins/plugin-splitview/src/index.ts
@@ -2,4 +2,5 @@
 // Copyright 2023 DXOS.org
 //
 
+export * from './SplitViewContext';
 export * from './SplitViewPlugin';
