+++ packages/apps/plugins/plugin-stack/src/StackPlugin.tsx
@@ -2,13 +2,14 @@
 // Copyright 2023 DXOS.org
 //
 
-import { definePlugin, PluginDefinition } from '@dxos/react-surface';
+import { TranslationsProvides } from '@braneframe/plugin-theme';
+import { PluginDefinition } from '@dxos/react-surface';
 
 import { StackMain } from './components';
 import { isStack } from './props';
 import translations from './translations';
 
-export const StackPlugin: PluginDefinition = definePlugin({
+export const StackPlugin = (): PluginDefinition<TranslationsProvides> => ({
   meta: {
     id: 'dxos:stack',
   },
