+++ packages/apps/plugins/plugin-theme/src/ThemePlugin.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { ThemeMode, ThemeProvider, Toast, Tooltip } from '@dxos/aurora';
 import { appTx } from '@dxos/aurora-theme';
 import { createStore } from '@dxos/observable-object';
-import { definePlugin, Plugin } from '@dxos/react-surface';
+import { Plugin, PluginDefinition } from '@dxos/react-surface';
 
 import compositeEnUs from './translations/en-US';
 
@@ -16,46 +16,46 @@ export type TranslationsProvides = {
   translations: Resource[];
 };
 
-type TranslationsPlugin = Plugin<TranslationsProvides>;
-
-const resources: Resource[] = [compositeEnUs];
+export type TranslationsPlugin = Plugin<TranslationsProvides>;
 
 export const translationsPlugins = (plugins: Plugin[]): TranslationsPlugin[] => {
   return (plugins as TranslationsPlugin[]).filter((p) => Array.isArray(p.provides?.translations));
 };
 
-const themeStore = createStore<{ themeMode: ThemeMode }>({ themeMode: 'dark' });
+export const ThemePlugin = (): PluginDefinition => {
+  let modeQuery: MediaQueryList | undefined;
+  const resources: Resource[] = [compositeEnUs];
+  const themeStore = createStore<{ themeMode: ThemeMode }>({ themeMode: 'dark' });
 
-const setTheme = ({ matches: prefersDark }: { matches?: boolean }) => {
-  document.documentElement.classList[prefersDark ? 'add' : 'remove']('dark');
-  themeStore.themeMode = prefersDark ? 'dark' : 'light';
+  const setTheme = ({ matches: prefersDark }: { matches?: boolean }) => {
+    document.documentElement.classList[prefersDark ? 'add' : 'remove']('dark');
+    themeStore.themeMode = prefersDark ? 'dark' : 'light';
+  };
+
+  return {
+    meta: {
+      id: 'dxos:ThemePlugin',
+    },
+    ready: async (plugins) => {
+      modeQuery = window.matchMedia('(prefers-color-scheme: dark)');
+      setTheme({ matches: modeQuery.matches });
+      modeQuery.addEventListener('change', setTheme);
+      for (const plugin of translationsPlugins(plugins)) {
+        resources.push(...plugin.provides.translations);
+      }
+    },
+    unload: async () => {
+      return modeQuery?.removeEventListener('change', setTheme);
+    },
+    provides: {
+      context: ({ children }) => (
+        <ThemeProvider {...{ tx: appTx, themeMode: themeStore.themeMode, resourceExtensions: resources }}>
+          <Toast.Provider>
+            <Tooltip.Provider>{children}</Tooltip.Provider>
+            <Toast.Viewport />
+          </Toast.Provider>
+        </ThemeProvider>
+      ),
+    },
+  };
 };
-
-let modeQuery: MediaQueryList | undefined;
-
-export const ThemePlugin = definePlugin({
-  meta: {
-    id: 'dxos:ThemePlugin',
-  },
-  ready: async (plugins) => {
-    modeQuery = window.matchMedia('(prefers-color-scheme: dark)');
-    setTheme({ matches: modeQuery.matches });
-    modeQuery.addEventListener('change', setTheme);
-    for (const plugin of translationsPlugins(plugins)) {
-      resources.push(...plugin.provides.translations);
-    }
-  },
-  unload: async () => {
-    return modeQuery?.removeEventListener('change', setTheme);
-  },
-  provides: {
-    context: ({ children }) => (
-      <ThemeProvider {...{ tx: appTx, themeMode: themeStore.themeMode, resourceExtensions: resources }}>
-        <Toast.Provider>
-          <Tooltip.Provider>{children}</Tooltip.Provider>
-          <Toast.Viewport />
-        </Toast.Provider>
-      </ThemeProvider>
-    ),
-  },
-});
