+++ packages/apps/plugins/plugin-treeview/src/components/index.ts
@@ -5,3 +5,4 @@
 export * from './BranchTreeItem';
 export * from './LeafTreeItem';
 export * from './TreeView';
+export * from './TreeViewContainer';
