+++ packages/apps/plugins/plugin-treeview/src/index.ts
@@ -2,4 +2,5 @@
 // Copyright 2023 DXOS.org
 //
 
+export * from './TreeViewContext';
 export * from './TreeViewPlugin';
